code fix to controller

+++ app/Http/Controllers/TasksController.php
@@ -16,11 +16,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $user = \Auth::user();
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::user()) {
+            $tasks = $user->tasks;
+
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        
+        return view('dashboard');
     }
 
     /**
@@ -53,6 +59,7 @@ class TasksController extends Controller
         $task = new Task;
         $task->content = $request->content;
         $task->status = $request->status;    // 追加したカラム
+        $task->user_id = $request->user()->id;
         $task->save();
 
         return redirect('/');
@@ -67,10 +74,9 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
-
+        
         return view('tasks.show', [
-            'task' => $task,
-        ]);
+        'task' => $task,]);
     }
 
     /**
@@ -81,11 +87,15 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = task::findOrFail($id);
+        $task = Task::findOrFail($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        
+        return view('dashboard');
     }
 
     /**
@@ -104,11 +114,15 @@ class TasksController extends Controller
         
         $task = Task::findOrFail($id);
 
-        $task->content = $request->content;
-        $task->status = $request->status;    // 追加したカラム
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            
+            return redirect('/');
+        }
 
-        return redirect('/');
+        return view('dashboard');
     }
 
     /**
@@ -121,8 +135,12 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
 
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            
+            return redirect('/');
+        }
 
-        return redirect('/');
+        return view('dashboard');
     }
 }
